Fungtion zum umrechnen von Maus to Sim. Positionen werden jetzt in Float angegeben.

+++ E-Feld-Simulation/E-Feld-Simulation/Punktladung.h
@@ -2,15 +2,16 @@
 #include <SFML\Graphics.hpp>
 
 #include "Math_PH.h"
+#include "InConvert.h"
 
 class Punktladung
 {
 public:
-	Punktladung(Physik_Value Ladung, sf::Vector2i position);
+	Punktladung(Physik_Value Ladung, sf::Vector2f position);
 	~Punktladung();
 
 	Physik_Value Q;
-	sf::Vector2i pos;
+	sf::Vector2f pos;
 	sf::Color col;
 	sf::CircleShape Erstelle_Kreis_Ladung(Physik_Value Radius);
 };
